pyproject.toml
[tool.poetry]
name = "httprunner"
version = "2.4.5"
description = "One-stop solution for HTTP(S) testing."
license = "Apache-2.0"
readme = "README.md"
authors = ["debugtalk <debugtalk@gmail.com>"]

homepage = "https://github.com/HttpRunner/HttpRunner"
repository = "https://github.com/HttpRunner/HttpRunner"
documentation = "https://docs.httprunner.org"

keywords = ["HTTP", "api", "test", "requests", "locustio"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

include = ["CHANGELOG.md", "httprunner/static/*"]

[tool.poetry.dependencies]
python = "~2.7 || ^3.5"
requests = "^2.22.0"
requests-toolbelt = "^0.9.1"
pyyaml = "^5.1.2"
jinja2 = "^2.10.3"
har2case = "^0.3.1"
colorama = "^0.4.1"
colorlog = "^4.0.2"
filetype = "^1.0.5"
jsonpath = "^0.82"
future = { version = "^0.18.1", python = "~2.7" }
sentry-sdk = "^0.13.5"

[tool.poetry.dev-dependencies]
flask = "<1.0.0"
coverage = "^4.5.4"

[tool.poetry.scripts]
hrun = "httprunner.cli:main"
ate = "httprunner.cli:main"
httprunner = "httprunner.cli:main"
locusts = "httprunner.plugins.locusts.cli:main"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
